+++ src/avatar/index.ts
@@ -1,2 +1,4 @@
 export * from "./Avatar";
+export * from "./avatarLoader";
+export * from "./avatarRadius";
 export * from "./avatarStyles";